requirements.txt
annotated-types==0.6.0
blinker==1.7.0
blis==0.7.11
bson==0.5.10
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
dnspython==2.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fsspec==2023.12.2
huggingface-hub==0.20.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.1.4
pillow==10.2.0
preshed==3.0.9
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==3.12.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
textblob==0.17.1
thinc==8.2.2
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
